+++ DatabaseProject/Data/Models/Review.cs
@@ -10,11 +10,11 @@ namespace Data.Models
     {
         public int ReviewID { get; set; }
         public int UserID { get; set; }
-        public int Rating { get; set; }
+        public decimal Rating { get; set; }
         public string Comment { get; set; }
         public int MovieID { get; set; }
 
-        public Review(int reviewID, int userID, int rating, string comment, int movieID)
+        public Review(int reviewID, int userID, decimal rating, string comment, int movieID)
         {
             ReviewID = reviewID;
             UserID = userID;
